requirements.txt
Flask==3.0.3
reportlab==4.2.2
supabase==2.5.0
python-dotenv==1.0.1
requests==2.32.3
